+++ files/gitlab-cookbooks/gitaly/recipes/enable.rb
@@ -46,7 +46,8 @@ node.default['gitaly']['env'] = {
   'PYTHONPATH' => "#{node['package']['install-dir']}/embedded/lib/python3.4/site-packages",
   # Charlock Holmes and libicu will report U_FILE_ACCESS_ERROR if this is not set to the right path
   # See https://gitlab.com/gitlab-org/gitlab-ce/issues/17415#note_13868167
-  'ICU_DATA' => "#{node['package']['install-dir']}/embedded/share/icu/current"
+  'ICU_DATA' => "#{node['package']['install-dir']}/embedded/share/icu/current",
+  'SSL_CERT_DIR' => "#{node['package']['install-dir']}/embedded/ssl/certs/"
 }
 
 env_dir env_directory do