+++ spec/chef/recipes/alertmanager_spec.rb
@@ -17,7 +17,7 @@ alertmanager_yml_output = <<-ALERTMANAGERYML
 ALERTMANAGERYML
 
 describe 'gitlab::alertmanager' do
-  let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(account)).converge('gitlab::default') }
+  let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(account env_dir)).converge('gitlab::default') }
 
   before do
     allow(Gitlab).to receive(:[]).and_call_original
@@ -43,6 +43,8 @@ describe 'gitlab::alertmanager' do
     end
 
     it_behaves_like 'enabled runit service', 'alertmanager', 'root', 'root'
+
+    it_behaves_like 'enabled env', '/opt/gitlab/etc/alertmanager/env', "SSL_CERT_DIR", '/opt/gitlab/embedded/ssl/certs/'
 
     it 'populates the files with expected configuration' do
       expect(config_template).to notify('ruby_block[reload alertmanager svlogd configuration]')
@@ -91,6 +93,9 @@ describe 'gitlab::alertmanager' do
         alertmanager: {
           listen_address: ':9093',
           enable: true,
+          env: {
+            'USER_SETTING' => 'asdf1234'
+          }
         },
         gitlab_rails: {
           smtp_enable: true,
@@ -114,5 +119,7 @@ describe 'gitlab::alertmanager' do
       expect(chef_run).to render_file('/var/opt/gitlab/alertmanager/alertmanager.yml')
         .with_content(/smtp_smarthost: other-testhost:465/)
     end
+
+    it_behaves_like 'enabled env', '/opt/gitlab/etc/alertmanager/env', "USER_SETTING", 'asdf1234'
   end
 end