+++ spec/chef/recipes/prometheus_spec.rb
@@ -169,7 +169,7 @@ prometheus_yml_output = <<-PROMYML
 PROMYML
 
 describe 'gitlab::prometheus' do
-  let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(account)).converge('gitlab::default') }
+  let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(account env_dir)).converge('gitlab::default') }
 
   before do
     allow(Gitlab).to receive(:[]).and_call_original
@@ -190,6 +190,8 @@ describe 'gitlab::prometheus' do
     end
 
     it_behaves_like 'enabled runit service', 'prometheus', 'root', 'root', 'gitlab-prometheus', 'gitlab-prometheus'
+
+    it_behaves_like 'enabled env', '/opt/gitlab/etc/prometheus/env', "SSL_CERT_DIR", '/opt/gitlab/embedded/ssl/certs/'
 
     it 'populates the files with expected configuration' do
       expect(config_template).to notify('ruby_block[reload prometheus svlogd configuration]')